+++ libs/rxjs/random/src/lib/from-crypto-random.ts
@@ -1,6 +1,11 @@
 import { Observable, Subscriber, timer } from 'rxjs';
 import { finalize, map, takeWhile, tap } from 'rxjs/operators';
 
+/**
+ * An Observable string value generator that generates random integers using
+ * the `Crypto.getRandomValues` method.
+ * @param emitDelay
+ */
 export function fromCryptoRandom(emitDelay = 0) {
   return new Observable((subscriber: Subscriber<number>) => {
     const arr = new Uint32Array(1);