+++ libs/rxjs/random/src/lib/from-random-string.ts
@@ -2,20 +2,16 @@ import { Observable, Subscriber, timer } from 'rxjs';
 import { finalize, map, takeWhile, tap } from 'rxjs/operators';
 
 /**
- * An Observable string value generator that generates random integers
- * @param min
- * @param max
+ * An Observable string value generator that generates random strings of the passed length
+ * @param length
  * @param emitDelay
  */
-export function fromRandomInt(min = 0, max = 1, emitDelay = 0) {
-  min = Math.ceil(min);
-  max = Math.floor(max);
-
-  return new Observable((subscriber: Subscriber<number>) => {
+export function fromRandomStr(length = 10, emitDelay = 0) {
+  return new Observable((subscriber: Subscriber<string>) => {
     timer(0, emitDelay)
       .pipe(
         takeWhile(() => !subscriber.closed),
-        map((value) => Math.floor(Math.random() * (max - min + 1) + min)),
+        map((value) => [...Array(length)].map(() => (~~(Math.random() * 36)).toString(36)).join('')),
         tap((value) => subscriber.next(value)),
         finalize(() => subscriber.complete()),
       )