+++ libs/rxjs/random/src/lib/from-random.ts
@@ -1,6 +1,12 @@
 import { Observable, Subscriber, timer } from 'rxjs';
 import { finalize, map, takeWhile, tap } from 'rxjs/operators';
 
+/**
+ * An Observable string value generator that generates random numbers
+ * @param min
+ * @param max
+ * @param emitDelay
+ */
 export function fromRandom(min = 0, max = 1, emitDelay = 0) {
   return new Observable((subscriber: Subscriber<number>) => {
     timer(0, emitDelay)